revert(goal): drop web UI + SSE event surface, keep slash-only
Following Codex's /goal model, the slash surface is the entire user-
facing API for goals. The web UI components (badge / dialog / hook),
chat-screen integration, and the supporting SSE event types
(goal_created / goal_status_changed / goal_iteration / goal_cleared)
were added in the previous commit but introduce more complexity than
they remove: inconsistent slash-vs-UI behaviour, audit-prompt
collapsibles that mostly confuse users, iterations-vs-rounds dual
counters, and a parallel emit pipeline (slashResult.events + REST
hub publishes) that only existed to feed the now-deleted UI.

Removed
- web/src/components/goal-badge.tsx
- web/src/components/goal-create-dialog.tsx
- web/src/hooks/use-goal.ts
- chat-screen.tsx goal integration (badge mount, dialog, 🎯 button,
  goal event consumption, goal-context collapsed note rendering,
  "goal-context" ChatMessage role, history goal-context branch)
- api.ts goal types (Goal / GoalStatus / CreateGoalRequest) and the
  five REST helpers; ChatStreamEvent goal_* variants and the origin
  / internal fields on ChatHistoryMessage
- Event* constants in agent/events.go
- slashResult.events plumbing
- agent/goal_events.go: event constructors + emitGoalIterationIfContinuation
  and goalToView helpers
- NewGoalToolEventHook and the AfterModelCall emitEvent on budget exhaustion
- handlers_goal.go publishGoalEvent + goalRecordView
- WebChatHistory's goal_context surfacing — back to hiding runtime-
  injected messages outright, matching Codex

Kept
- REST endpoints under /api/agents/{id}/goal — invisible to chat
  users, useful for external automation. No SSE pair anymore.
- Routing tuple validation on REST POST create (channel + chatId
  required) plus the matching test, since the silent-failure mode
  it closes is real regardless of UI presence.
- Plan-mode precedence over goal (auto-plan skip + explicit plan
  flag stripped when a goal is active). sessionHasActiveGoal helper
  moved into loop.go since goal_events.go is gone.
